Refactor `BookManager` into standalone module, streamline `trade_listener_channels.rs`, enhance imports via `prelude`, and bump version to `0.4.3`.
