+++ Makefile
@@ -153,7 +153,7 @@ bench-show:
 
 .PHONY: bench-save
 bench-save: check-cargo-criterion
-	cargo criterion --output-format quiet --history-id v0.4.1 --history-description "Version 0.3.2 baseline"
+	cargo criterion --output-format quiet --history-id v0.4.3 --history-description "Version 0.3.2 baseline"
 
 .PHONY: bench-compare
 bench-compare: check-cargo-criterion
